+++ packages/twopoint5d-testing/test/number-or-the-beast.test.js
@@ -1,7 +1,0 @@
-import { expect } from '@esm-bundle/chai';
-
-describe('number of the beast', () => {
-  it('the number of the beast is the number of the beast', () => {
-    expect(666).to.equal(666);
-  });
-});